--- a/documentation/docs/fundamentals/defining-your-theme.md
+++ b/documentation/docs/fundamentals/defining-your-theme.md
@@ -34,10 +34,6 @@ const theme = createTheme({
     l: 24,
     xl: 40,
   },
-  breakpoints: {
-    phone: 0,
-    tablet: 768,
-  },
 });
 
 export type Theme = typeof theme;