--- a/src/composeRestyleFunctions.ts
+++ b/src/composeRestyleFunctions.ts
@@ -1,4 +1,4 @@
-import {StyleSheet} from 'react-native';
+import {StyleSheet, ViewStyle} from 'react-native';
 
 import {
   RestyleFunctionContainer,
@@ -46,16 +46,21 @@ const composeRestyleFunctions = <
       dimensions,
     }: {
       theme: Theme;
-      dimensions: Dimensions;
+      dimensions: Dimensions | null;
     },
   ): RNStyle => {
-    const styles = Object.keys(props).reduce(
-      (styleObj, propKey) => ({
-        ...styleObj,
-        ...funcsMap[propKey as keyof TProps](props, {theme, dimensions}),
-      }),
-      {},
-    );
+    const styles: ViewStyle = {};
+    const options = {theme, dimensions};
+    // We make the assumption that the props object won't have extra prototype keys.
+    // eslint-disable-next-line guard-for-in
+    for (const key in props) {
+      const mappedProps = funcsMap[key](props, options);
+      // eslint-disable-next-line guard-for-in
+      for (const mappedKey in mappedProps) {
+        styles[mappedKey as keyof ViewStyle] = mappedProps[mappedKey];
+      }
+    }
+
     const {stylesheet} = StyleSheet.create({stylesheet: styles});
     return stylesheet;
   };