--- a/src/hooks/useRestyle.ts
+++ b/src/hooks/useRestyle.ts
@@ -2,7 +2,6 @@ import {useMemo} from 'react';
 import {StyleProp, useWindowDimensions} from 'react-native';
 
 import {BaseTheme, RNStyle, Dimensions} from '../types';
-import {getKeys} from '../typeHelpers';
 
 import useTheme from './useTheme';
 
@@ -13,33 +12,23 @@ const filterRestyleProps = <
   componentProps: TProps,
   omitPropertiesMap: {[key in keyof TProps]: boolean},
 ) => {
-  const props = omitPropertiesMap.variant
-    ? {variant: 'defaults', ...componentProps}
-    : componentProps;
-  return getKeys(props).reduce(
-    ({cleanProps, restyleProps, serializedRestyleProps}, key) => {
-      if (omitPropertiesMap[key as keyof TProps]) {
-        return {
-          cleanProps,
-          restyleProps: {...restyleProps, [key]: props[key]},
-          serializedRestyleProps: `${serializedRestyleProps}${String(key)}:${
-            props[key]
-          };`,
-        };
-      } else {
-        return {
-          cleanProps: {...cleanProps, [key]: props[key]},
-          restyleProps,
-          serializedRestyleProps,
-        };
-      }
-    },
-    {cleanProps: {}, restyleProps: {}, serializedRestyleProps: ''} as {
-      cleanProps: TProps;
-      restyleProps: TRestyleProps;
-      serializedRestyleProps: string;
-    },
-  );
+  const cleanProps: TProps = {} as TProps;
+  const restyleProps: TProps & {variant?: unknown} = {} as TProps;
+  let serializedRestyleProps = '';
+  if (omitPropertiesMap.variant) {
+    restyleProps.variant = componentProps.variant ?? 'defaults';
+  }
+  for (const key in componentProps) {
+    if (omitPropertiesMap[key as keyof TProps]) {
+      restyleProps[key] = componentProps[key];
+      serializedRestyleProps += `${String(key)}:${componentProps[key]};`;
+    } else {
+      cleanProps[key] = componentProps[key];
+    }
+  }
+
+  const keys = {cleanProps, restyleProps, serializedRestyleProps};
+  return keys;
 };
 
 const useRestyle = <
@@ -55,7 +44,7 @@ const useRestyle = <
         dimensions,
       }: {
         theme: Theme;
-        dimensions: Dimensions;
+        dimensions: Dimensions | null;
       },
     ) => RNStyle;
     properties: (keyof TProps)[];
@@ -64,22 +53,30 @@ const useRestyle = <
   props: TProps,
 ) => {
   const theme = useTheme<Theme>();
-  const dimensions = useWindowDimensions();
+
+  // Theme should not change between renders, so we can disable rules-of-hooks
+  // We want to avoid calling useWindowDimensions if breakpoints are not defined
+  // as this hook is called extremely often and incurs some performance hit.
+  const dimensions = theme.breakpoints
+    ? // eslint-disable-next-line react-hooks/rules-of-hooks
+      useWindowDimensions()
+    : null;
 
   const {cleanProps, restyleProps, serializedRestyleProps} = filterRestyleProps(
     props,
     composedRestyleFunction.propertiesMap,
   );
 
-  const calculatedStyle = useMemo(() => {
+  const calculatedStyle: StyleProp<RNStyle> = useMemo(() => {
     const style = composedRestyleFunction.buildStyle(restyleProps as TProps, {
       theme,
       dimensions,
     });
 
-    const styleProp = props.style;
+    const styleProp: StyleProp<RNStyle> = props.style;
     if (typeof styleProp === 'function') {
-      return (...args: any[]) => [style, styleProp(...args)].filter(Boolean);
+      return ((...args: any[]) =>
+        [style, styleProp(...args)].filter(Boolean)) as StyleProp<RNStyle>;
     }
     return [style, styleProp].filter(Boolean);
 
@@ -95,10 +92,8 @@ const useRestyle = <
     composedRestyleFunction,
   ]);
 
-  return {
-    ...cleanProps,
-    style: calculatedStyle,
-  };
+  cleanProps.style = calculatedStyle;
+  return cleanProps;
 };
 
 export default useRestyle;