--- a/src/utilities/index.ts
+++ b/src/utilities/index.ts
@@ -1,0 +1,4 @@
+export {getValueForScreenSize} from './getValueForScreenSize';
+export {getResponsiveValue} from './getResponsiveValue';
+export {isResponsiveObjectValue} from './isResponsiveObjectValue';
+export {getThemeValue} from './getThemeValue';