## Description

`restyle` is in its essence a runtime library to achieve better DX. Ideally, `restyle` should have _no_ overhead and _no_ performance implications. Since, as things stand, we _need_ to run code in the runtime to convert theme values to RN style, there will be some overhead.

However, we should ensure that the overhead is as small as possible. We've already made some performance improvements [in the past](https://github.com/Shopify/restyle/commit/cdf1520943c7e8af8762d0cae33e88c9760e60c6) but we still found there's a noticeable performance drop when using restyle.

Using the fixture inside the `flash-list` repo in [this branch](https://github.com/Shopify/flash-list/tree/fixture/restyle), we found out `restyle` still incurs performance drop. Measuring on iPad Pro 2019, we found the FPS drops by **~7 %** and list load time takes **12 %** longer. Considering these numbers, we decided to investigate further why `restyle` still has such an impact on performance.

This PR includes multiple small performance improvements but the major ones have been:
- Migrating from `reduce` to `for-loop` in `filterRestyleProps`. Creating many objects on every `useRestyle` render is _expensive_. `for-loop` with updating objects in-place instead of copying and creating new ones is _much_ more efficient.
- `breakpoints` are now optional. Querying window dimensions and computing responsive values has additional performance cost. If developers are not using `breakpoints`, we should not compute these values. In the future, we should consider creating responsive variants of restyle components, so the responsive computations can be done really only for those components that need them.
- Style memoization. We currently have `useMemo` inside `useRestyle` to ensure a component does not need to recompute its style on every render. We decided to take this step further and memoize computed style inside the `theme` object. This means a style does not have to be recomputed if any other component has already computed the same theme value.

## Measurements

All the following measurements have been done in [this](https://github.com/Shopify/flash-list/tree/fixture/restyle) flash-list branch on iPad Pro 2019:

#### FPS

Without restyle:
118.2, 116, 116.6, 117.1, 116.7, 115.7, 116.2, 115.7, 116.1, 115.3, 115.9
Avg: 116.32

With restyle:
109.7, 108.3, 106.8, 108.5, 110.4, 110.2, 108.0, 108.8, 107.0, 110.1
Avg: 108.78
**Perf drop: 7 %**

With this PR's optimizations:
115.7, 115.2, 116.1, 113.8, 115.1, 115.4, 115.0, 114.7, 111.1, 113.8, 113.1
Avg: 114.45
**Perf drop: 1.5 %**
**Relative perf improvement: 79 %**

#### List load time

Without restyle:
61, 60, 62, 59, 61, 59, 61, 64, 57, 61
Avg: 60.5

With restyle:
71, 67, 71, 69, 70, 64, 79, 61, 71, 65
Avg: 68.8
**Performance drop: 12 %**

With optimizations:
64, 65, 66, 60, 64, 64, 67, 63, 66, 64
Average: 64.3
Performance drop: 6 %
**Relative performance improvement: 50 %**

#### Tracer

We have created [this](https://github.com/Shopify/restyle/blob/performance/tracer/src/tracer.ts) helper class called `tracer` to get granular numbers on how much time JS spends on evaluating `useRestyle`

**Time spent on useRestyle in ms for ~30000 calls**:
Without optimizations:
563, 533, 515, 618, 583
Average: 562.4 ms

With optimizations:
270, 253, 242, 217, 262
Average: 248.8 ms
**Performance improvement: 56 %**

From this, more than 50 % is spent on `filterRestyleProps` method. I don't see how we can effectively get rid of that for loop since we need it to computed the `serializedRestyleProps` that determine whether a style needs to be recomputed or not.

## Summary

The changes in this PR lead to large performance improvements where the FPS when scrolling in `flash-list` **has improved by almost 80 %**. In the FlashList fixture, using restyle still incurs ~1.5 % FPS drop. We expect this number to be smaller in more complicated lists as restyle would constitute smaller part of all the work done on every render.

The total time spent on `useRestyle` has dropped by more than 50 %, the list load time has improved by 50 %.

There is still some performance drop attached to `useRestyle` even with these new changes but the impact should be now minimal.

<!--
Please include a summary of what you want to achieve in this pull request. Remember to indicate the affected package(s).
-->

## Reviewers’ hat-rack :tophat:

- Check out fixture, everything should still work as expected.
- If you are interested, additional measurements would be much appreciated. Let me know and I can give you some additional guidance how to do so!

## Checklist

- [ ] I have added a changelog entry following the [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) guidelines.