@craftzdog 2023-06-08
Great work @fortmarek!
Can `useResponsiveProp` also be improved as it uses `useWindowDimensions`?
It affects a lot when the device orientation is changed on tablets.